requirements.txt
raccoontools==1.0.0
pillow==10.4.0
pydantic==2.6.4
requests==2.32.3
python-dotenv==1.0.0